third_party/WebKit/LayoutTests/virtual/outofblink-cors-ns/http/tests/fetch/chromium/data-saver-expected.txt

This is a testharness.js-based test.
PASS fetch() from document with Data-Saver disabled.
PASS fetch() from dedicated worker with Data-Saver disabled.
PASS fetch() from shared worker with Data-Saver disabled.
PASS fetch() from document with Data-Saver enabled.
PASS fetch() from dedicated worker with Data-Saver enabled.
FAIL fetch() from shared worker with Data-Saver enabled. promise_test: Unhandled rejection with value: object "Error: assert_equals: Save-Data header should be sent for worker script when enabled. expected "Save-Data: on" but got "No Save-Data""
Harness: the test ran to completion.
